+++ src/components/FormConstructor/FormConstructorFormBlockConfigure/Panels/Settings/TextSettings/types.ts
